refactor: move AgendaScreenState to separate file
